move AlertDaoTest to new module
